Reestructure files and folders
