+++ provers/stark/Cargo.toml
@@ -13,7 +13,7 @@ lambdaworks-math = { workspace = true , features = ["lambdaworks-serde"] }
 lambdaworks-crypto.workspace = true
 winterfell = { git = "https://github.com/facebook/winterfell" }
 winter-utils = { git = "https://github.com/facebook/winterfell" }
-
+examples = { git = "https://github.com/facebook/winterfell" }
 rand = "0.8.5"
 thiserror = "1.0.38"
 log = "0.4.17"
