+++ provers/stark/src/lib.rs
@@ -18,7 +18,7 @@ pub mod traits;
 pub mod transcript;
 pub mod utils;
 pub mod verifier;
-pub mod winterfell;
+pub mod winterfell_adapter;
 
 #[cfg(test)]
 pub mod tests;
